Changed the basis for interpolation to allow using RegularGridInterpolator
Previous versions of the code did not take advantage of how much simpler it is to build the interpolator object on a regular lat/lon/depth grid.

Instead of :
Projection to UTM > Interpolation
the code now performs:
Interpolation > Projection to UTM

This should make the 3D interpolation much faster.

Also added a basic chunking of the interpolation queries, to allow for progress to be printed, and enable easy parallelization in the future