Add auto-smoothing info print to the tomography processor.
I've made changes so that the topo processor now logs values of the iterative smoothing process, so that we can print out the parameters in the client logs.
It also prints the equivalent standard deviation (in pixel and meters) that we would need to use if we used the linear gaussian filter instead of the iterative-diffusion process.

Also updated the docstrings for the `auto_smooth_topography` function to Numpy-style standard.